+++ testdata/engineexport/lab/issue-35-oie-entrypoint.sh
@@ -1,0 +1,6 @@
+#!/bin/sh
+set -eu
+mkdir -p /lab
+tar -xzf /opt/oie.tar.gz -C /lab
+cd /lab/oie
+exec ./oieserver